Expose addComputedDependency so computeds can change their dependencies (#7)
I'm in a situation where i have a computed observable that depends on a
dynamic list of other observables (things can be added to the list
throughout the lifetime of the computed). I order to support that I need
access to the addComputedDependency function, which I can use as is
without any changes.